group `ty::opaque` in one file
