+++ compiler/rustc_middle/src/ty/mod.rs
@@ -78,9 +78,7 @@ mod infer_var_info;
 mod instance;
 mod list;
 mod main_definition;
-mod opaque_hidden_type;
-mod opaque_type_key;
-mod opaque_types;
+mod opaque;
 mod param_env;
 mod parameterized;
 mod placeholder;
@@ -145,8 +143,7 @@ pub use self::infer_var_info::InferVarInfo;
 pub use self::instance::{Instance, InstanceDef, ShortInstance, UnusedGenericParams};
 pub use self::list::List;
 pub use self::main_definition::MainDefinition;
-pub use self::opaque_hidden_type::OpaqueHiddenType;
-pub use self::opaque_type_key::OpaqueTypeKey;
+pub use self::opaque::{OpaqueHiddenType, OpaqueTypeKey};
 pub use self::param_env::{ParamEnv, ParamEnvAnd};
 pub use self::parameterized::ParameterizedOverTcx;
 pub use self::placeholder::{Placeholder, PlaceholderConst, PlaceholderRegion, PlaceholderType};
